Revert "refactor: fold the site Camera type into data-renderers/types"
This reverts commit 929ae56afa38ba5b7b8be0287c2ce02877e3fc45.
