+++ site/src/examples/data-renderers/types.ts
@@ -3,13 +3,6 @@
 import type { Interval, box2D, vec2, vec4 } from '@alleninstitute/vis-geometry';
 import type { NormalStatus } from '@alleninstitute/vis-core';
 
-// a basic camera, for viewing slices
-export type Camera = {
-    readonly view: box2D; // a view in 'data space'
-    readonly screen: vec2; // what that view projects to in display space, aka pixels
-    readonly projection: 'webImage' | 'cartesian';
-};
-
 type ChannelSettings = {
     gamut: Interval;
     index: number;
