+++ site/src/examples/data-renderers/volumeSliceRenderer.ts
@@ -1,9 +1,10 @@
 import { type AsyncDataCache, beginLongRunningFrame, logger } from '@alleninstitute/vis-core';
 import type REGL from 'regl';
-import type { Camera, RenderCallback } from './types';
+import type { RenderCallback } from './types';
 
 import { Box2D, CartesianPlane, Vec2, type vec2 } from '@alleninstitute/vis-geometry';
 import { pickBestScale, sizeInUnits, sizeInVoxels } from '@alleninstitute/vis-omezarr';
+import type { Camera } from '../common/camera';
 import type { AxisAlignedZarrSlice } from '../data-sources/ome-zarr/planar-slice';
 import type { AxisAlignedZarrSliceGrid } from '../data-sources/ome-zarr/slice-grid';
 import { applyOptionalTrn } from './utils';
